+++ changelog.md
@@ -16,6 +16,7 @@ Codion Change Log
 - FilterTable.model() added.
 - Control related renaming and refactoring.
 - ControlId, ControlShortcuts and ControlSet added, KeyboardShortcuts removed, related refactoring.
+- KeyEvents, special handling added for JSpinner.
 ### is.codion.swing.framework.model
 - SwingEntityTableModel.EntityTableColumns, problematic override of comparable() removed.
 ### is.codion.swing.framework.model.tools